requirements.txt
google_api_python_client==2.74.0
hikari==2.0.0.dev115
hikari_lightbulb==2.3.1
lavasnek_rs==0.1.0a5
python-dotenv==0.21.1
requests==2.27.1
